corna->meta.py: add meta utils
There are utility functions that use functionality in other modules in the
`utils` directory. This has led to some annoying circular dependencies.

This commit creates a `meta` utility which houses functionality shared in the
`utils` directory itself as well as anywhere else in the repo.
