+++ corna/controls/auth_control.py
@@ -3,6 +3,7 @@ import logging
 from typing import Any, Dict, Optional
 
 from corna.db import models
+from corna.utils import get_utc_now
 from corna.utils import secure, utils
 from corna.utils.errors import (
     IncorrectPasswordError, NoneExistingUserError, UserExistsError)
