+++ corna/controls/post_control.py
@@ -13,6 +13,7 @@ from werkzeug.utils import secure_filename
 
 from corna.db import models
 from corna.enums import ContentType
+from corna.utils import get_utc_now
 from corna.utils import image_proc, secure, utils
 from corna.utils.utils import current_user
 from corna.utils.errors import (
