+++ corna/utils/__init__.py
@@ -1,4 +1,12 @@
 from .vault_manager import get_item as vault_item
 from .vault_manager import get_items as vault_items
 
-__all__ = ["vault_item", "vault_items"]
+# this is being done to avoid circular dependencies
+from .meta import future, get_utc_now
+
+__all__ = [
+    "future",
+    "get_utc_now",
+    "vault_item",
+    "vault_items"
+]
