+++ corna/utils/utils.py
@@ -35,27 +35,6 @@ def respond_json_error(message: str, code: int) -> None:
     flask.abort(flask.make_response(flask.jsonify({'message': message}), code))
 
 
-def get_utc_now() -> datetime.datetime:
-    """Get a time-zone aware datetime for "now".
-
-    :returns: "now" as a datetime
-    :rtype: datetime.datetime
-    """
-    return datetime.datetime.now(datetime.timezone.utc)
-
-
-def future(days: int = 14) -> datetime.datetime:
-    """Datetime object for some number of days in the future.
-
-    :param int days: number of days in the future
-    :returns: current time + number of skip days
-    :rtype: datetime.datetime
-    """
-    now = get_utc_now()
-    skip = datetime.timedelta(days=days)
-    return now + skip
-
-
 def get_uuid() -> str:
     """Get a random UUID.
